Change Logs to Easy Travel

+++ src/main/java/team/easytravel/logic/commands/fixedexpense/AddFixedExpenseCommand.java
@@ -71,7 +71,7 @@ public class AddFixedExpenseCommand extends Command {
             model.addFixedExpense(toAdd);
         }
 
-        double currentBudget = model.getBudget();
+        int currentBudget = model.getBudget();
 
 
         if (currentBudget < 1) {
